ppoll, syscall for arm/mips and some minor fixes

+++ include/sys/syscall.h
@@ -3,6 +3,6 @@
 
 #include <asm/unistd.h>
 
-int syscall(int number, ...);
+long syscall(long number, ...);
 
 #endif
